SenSa/SenSa/requirements.txt
asgiref==3.11.1
attrs==26.1.0
autobahn==23.6.2
automat==25.4.16
black==25.11.0
cffi==2.0.0
cfgv==3.4.0
channels==4.1.0
channels-redis==4.2.0
click==8.1.8
constantly==23.10.4
cryptography==46.0.7
daphne==4.1.2
distlib==0.4.0
django==4.2.29
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
filelock==3.24.2
hyperlink==21.0.0
identify==2.6.15
idna==3.11
incremental==24.11.0
msgpack==1.1.2
mypy-extensions==1.1.0
nodeenv==1.10.0
packaging==26.0
pathspec==1.0.4
pillow==12.2.0
pip==26.0.1
platformdirs==4.4.0
pre-commit==4.3.0
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pyjwt==2.12.1
pyopenssl==26.0.0
python-discovery==1.2.0
python-dotenv==1.2.2
pytokens==0.4.1
pyyaml==6.0.3
redis==7.4.0
ruff==0.15.6
service-identity==24.2.0
setuptools==58.0.4
sqlparse==0.5.5
tomli==2.4.0
twisted==25.5.0
txaio==25.12.2
typing-extensions==4.15.0
virtualenv==21.2.0
zope-interface==8.3
